Added forum pages
This will help reduce the bandwith needed to serve forum posts to users, since it wil only send a handful at a time, instead of all of them.

Mainly an issue with lots (1000) forum posts, where it will take a good couple of seconds to load the page (and thoroughly destroying the server).